feat(@angular-devkit/core): move findTypes in utility and export
